# vulkan: conv2d addressing optimizations