@jeffbolznv 2025-08-03
I was surprised moving the invariants out of the loop helped, but it seemed to. Putting the load in the branch generates much better code for NVIDIA, doing a predicated load rather than the load + select.