@daniandtheweb 2025-08-03
I did a similar change by moving invariants outside the loops some time ago to the im2col shader and there was a measurable improvement on my AMD gpus so I guess the compiler doesn't optimize that automatically.